Introduce zone contexts
In rule definitions, `context` and `isContext` are replaced by
`requiresContext` and `definesContext`, respectively:
 - `requiresContext` accepts a string made of rule names and the usual
&, ! and | operators.
 - `definesContext` accepts `False`, `pageId`, `pageType` or `zone`

`pageId` and `pageType` both match (or discard) the whole document in
which they are found.
`zone` matches (or discards) the corresponding node and its sub-tree.

Additionally, `nodeManager.searchNode` is now aware of `multiple` and
`index` rule attributes in order to avoid useless excessive searches.