Rules: Introduce rule priority.
This implementation is a workarround for the lack of multi criteria sets rules
(planned feature): If sevaral rules share the same name and have each an
assigned priority, they are evaluated lowest priority first. As soon as results
are found for one of them, those with higher priorities are not evaluated.