Check Python web modules API compatibility.
New attribute `webModuleHandler.WebModule.API_VERSION`, first set to "0.1".

The Python module shipped with web modules should now provide a root attribute
`API_VERSION`, an `str` to be parsed by `packaging.version` and compared
against the current value defined in WebAccess.
As the API is not stable yet, any version that does not match exactly will be
rejected and the Python web module will not be loaded.
In the future, we expect the versioning scheme to adhere to semantic versioning
that is, starting with API version 1.0, Python web module will be loaded as
long as they conform to the sole major version - not only the exact same full
version.

Note: Python web modules are currently only shipped by integrators fully
handling the update process of the target corporate environment. They are
mainly intended to temporarily circumvent WebAccess limitations while the
missing functionalities keep being added to the core. They are not expected to
ever be available through the upcoming web modules cloud catalog, which should
provide JSON web modules only.