Rule Wizard
This first implementation mostly presents the main panels of the Rule Editor
in simple mode.
Hitting F12 brings up the editor pre-filled with the data already entered on
the wizard, with the corresponding category selected.
Hitting control+enter finishes the wizard, despite its lack for a finish button
before the last page.
The wizard is invoked by the "New Rule" WebAccess Menu entry as well as the new
and edit buttons on the Rule Manager.