Criteria Editor: testCriteria: Drop properties
For the test, the "multiple" property is forced to ensure the user is informed
about all potential matches. Hence, the rule type "marker" is forced as the
"multiple" property is not supported for every rule types.
Rather than filtering out properties that are not supported for markers, simply
drop them all as they have no impact on result matching anyway.