Multiple: Consider all alternatives
This change is only effective when the property is defined
at Rule level. It has no effect when defined at Criteria
level. Hence, to obtain the former behavior, which only
considers the first matching alternative, the property must
be set to False on the Rule and True on each of its alternative
Criteria.