Browsable Element Description dialog
Step towards #40

The report now presents a single node at a time, rather than the full branch
from the caret to the root.
All properties are always displayed, even when there is no associated value.
Typing a character with no modifier moves the caret to the line starting with that character.
The first mention is now the matching rules at the given node, which includes
results from all sub-modules and specifies start/end of free zones.
When inspecting the caret node, if it matches a rule, the text leaf is shown,
otherwise its parent.

The dialog now supports the following commands:
 - alt+u/d/l/r/a/b or alt+up/down/left/right/pgup/pgdn to walk nodes
 - (alt+)backspace to rollback history
 - alt+home to inspect the node for which the dialog was opened
 - alt+end to inspect the document caret node (currently the same as above)
 - F9 to mark a node
 - shift+F9 to inspect the marked node
 - F10 to compute the relative path from the marked node to the current node
   and copy it in the clipboard
 - enter to move the document caret to the position of the currently inspected node
 - control+V or shift+insert to walk the relative path found in the clipboard

These commands are also available through a context menu advertising the corresponding keyboard shortcuts.

The operations (except move/retrieve caret) are performed on a snapshot of the
document tree to circumvaint document refresh/reload.

Additionally:
 - Rename `elementDescription` to `inspector`
 - GUI: Editors: F8 opens the Element Description dialog