Clipboard: Copy/paste Rules/Alternatives/Selectors between WebModules
Copy using control+C or control+insert, paste using control+V or
shift+insert when focus is on:
 - Rules Manager: Rules/results tree
 - Rule Editor:
   - The criteria category on the categories tree
   - In full mode: The alternatives list on the criteria panel
 - Criteria Editor for a Free Zone, the start and end criteria
   categories on the categories list.

Copying stores the JSON representation of the rule or criteria set in
the clipboard in raw text mode.
A rule with a single alternative can be pasted in lieu of an alternative or selector.
An alternative with a single selector can be pasted in lieu of a selector.
An alternative or selector can be pasted as base for a new rule.
Pasting opens the corresponding wizard or editor to allow for alteration before committing.

 - Editors: Support shift+F6 (was F6 only)
 - Rule Editor: Alternatives List: Support deleting alternatives using the delete key